+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Change Log
 
+## v2.3.0
+  * PonyVCO
+    * Initial release
+  * EvenVCO
+    * Optionally remove DC from pulse wave output
+  * StereoStrip
+    * Address high CPU usage when using EQ sliders
+
 ## v2.2.0
 
   * StereoStrip